[add] parse files and headers

+++ include/color.h
@@ -6,7 +6,7 @@
 /*   By: jaejilee <jaejilee@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/17 17:35:45 by jaejilee          #+#    #+#             */
-/*   Updated: 2024/02/17 17:37:12 by jaejilee         ###   ########.fr       */
+/*   Updated: 2024/02/19 16:59:11 by jaejilee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ typedef struct s_color
 	int		red;
 	int		green;
 	int		blue;
-	float	brightness;
+	double	brightness;
 }	t_color;
 
 #endif
